+++ android/app/src/main/java/edu/illinois/rokwire/maps/MapView.java
@@ -402,7 +402,7 @@ public class MapView extends FrameLayout implements OnMapReadyCallback {
             
             JSONObject jsonLocation = new JSONObject();
             jsonLocation.put("latitude", latLng.latitude);
-            jsonLocation.put("longitude", latLng.latitude);
+            jsonLocation.put("longitude", latLng.longitude);
             jsonArgs.put("location", jsonLocation);
         } catch (JSONException e) {
             e.printStackTrace();
@@ -422,7 +422,7 @@ public class MapView extends FrameLayout implements OnMapReadyCallback {
 
             JSONObject jsonLocation = new JSONObject();
             jsonLocation.put("latitude", poi.latLng.latitude);
-            jsonLocation.put("longitude", poi.latLng.latitude);
+            jsonLocation.put("longitude", poi.latLng.longitude);
             jsonPOI.put("location", jsonLocation);
             jsonArgs.put("poi", jsonPOI);
 
@@ -430,7 +430,7 @@ public class MapView extends FrameLayout implements OnMapReadyCallback {
             e.printStackTrace();
         }
         String methodArguments = jsonArgs.toString();
-        MainActivity.invokeFlutterMethod("map.location.select", methodArguments);
+        MainActivity.invokeFlutterMethod("map.poi.select", methodArguments);
     }
 
     private void onCameraIdle() {